+++ src/Glpi/Helpdesk/Tile/FormTile.php
@@ -32,15 +32,39 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Form\SelfService;
+namespace Glpi\Helpdesk\Tile;
 
-interface TileInterface
+use Glpi\Form\Form;
+
+final class FormTile implements TileInterface
 {
-    public function getTitle(): string;
+    public function __construct(
+        private Form $form,
+        // The following properties might be specified to override the form's
+        // fields if needed.
+        private string $title = "",
+        private string $description = "",
+        private string $illustration = "",
+    ) {
+    }
 
-    public function getDescription(): string;
+    public function getTitle(): string
+    {
+        return $this->title ?: $this->form->fields['name'];
+    }
 
-    public function getIllustration(): string;
+    public function getDescription(): string
+    {
+        return $this->description ?: $this->form->fields['description'];
+    }
 
-    public function getLink(): string;
+    public function getIllustration(): string
+    {
+        return $this->illustration ?: $this->form->fields['illustration'];
+    }
+
+    public function getLink(): string
+    {
+        return '/Form/Render/' .  $this->form->getID();
+    }
 }
