+++ src/Toolbox.php
@@ -37,7 +37,7 @@ use Glpi\Application\View\TemplateRenderer;
 use Glpi\Console\Application;
 use Glpi\DBAL\QueryParam;
 use Glpi\Event;
-use Glpi\Form\DefaultFormsManager;
+use Glpi\Helpdesk\DefaultDataManager;
 use Glpi\Http\Response;
 use Glpi\Mail\Protocol\ProtocolInterface;
 use Glpi\Rules\RulesManager;
@@ -2157,8 +2157,8 @@ class Toolbox
             }
 
             // Create default forms
-            $default_forms_manager = new DefaultFormsManager();
-            $default_forms_manager->createDefaultForms();
+            $default_forms_manager = new DefaultDataManager();
+            $default_forms_manager->initializeData();
 
             // Initalize rules
             RulesManager::initializeRules();
